[CHANGE] Template tag library renamed
