+++ fleetpings/templates/fleetpings/bundles/fleetpings-css.html
@@ -1,3 +1,3 @@
-{% load fleetpings_versioned_static %}
+{% load fleetpings %}
 
 <link rel="stylesheet" href="{% fleetpings_static 'fleetpings/css/fleetpings.min.css' %}">
