+++ fleetpings/templates/fleetpings/bundles/fleetpings-js.html
@@ -1,3 +1,3 @@
-{% load fleetpings_versioned_static %}
+{% load fleetpings %}
 
 <script src="{% fleetpings_static 'fleetpings/js/fleetpings.min.js' %}"></script>
