+++ fleetpings/templates/fleetpings/index.html
@@ -1,7 +1,7 @@
 {% extends "fleetpings/base.html" %}
 
 {% load i18n %}
-{% load fleetpings_versioned_static %}
+{% load fleetpings %}
 
 {% block aa_fleetpings_body %}
     <div class="row">
