+++ fleetpings/templates/fleetpings/partials/form/segments/fleet-doctrine.html
@@ -1,4 +1,4 @@
-{% load fleetpings_urls %}
+{% load fleetpings %}
 
 {% if doctrines|length > 0 %}
     <datalist id="fleetDoctrineList">
